Add Button component
This adds a `<Button />` component, which roughly replaces the CSS for
`btn-cta`. It has a mandatory prop of `theme`, which currently only
supports `cta`.

If supplied a `to` prop, it is assumed to be a react-router link and the
rendered component is therefore a `Link`. Otherwise, the rendered
component is a standard `<button />` element.

As part of a light refactor, I've removed the dependency on an SVG file
by using a CSS clip-path to produce the "arrow" shape on the far right
of the button.